Expand supported versions of doctrine/annotations (#39)
This now supports v1 or v2 of the annotations package. v2 has dropped
support for SimpleAnnotationReader; consequently, a bunch of the tests
needed to be updated with different mappings to handle only the "full"
annotation reader being present in that version.

The default driver now switches based on class presence, still generally
matching Doctrine itself. This should not be a BC break in practice
since updating to that version would have already forced an eqiuivalent
model update.